Rename translate HOC to withNamespaces
