+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-i18next-phraseapp",
-  "version": "1.1.0",
+  "version": "2.0.0",
   "description": "Connect react-i18next with the PhraseApp In-Context Editor",
   "homepage": "https://phraseapp.com",
   "main": "./dist/react-i18next-phraseapp.js",
@@ -56,7 +56,7 @@
   "peerDependencies": {
     "i18next": ">=7.1.3",
     "react": "^15.4.2",
-    "react-i18next": ">=2.2.3"
+    "react-i18next": ">=8.0.0"
   },
   "files": [
     "dist"
